requirements.txt
fastapi
uvicorn
langchain_community
PyMuPDF
python-docx
pandas
markdown
unstructured
pytesseract
SQLAlchemy
psycopg2-binary
pgvector
torch
transformers
sentence-transformers
accelerate
bitsandbytes
psutil
python-dotenv 